commit Adriano #46 fine backend in FragmentMarangicoin.java cambiato fragment_layout_marangicoin da ConstraintLayout a RelativeLayout

+++ app/src/main/java/com/QUeM/TreGStore/DatabaseClass/Conti.java
@@ -33,7 +33,7 @@ public class Conti {
         this.coinAmount=newCoinAmount;
     }
 
-    public void setCurrentDiscount(int newCurrentDiscount){
+    public void setCurrentDiscount(double newCurrentDiscount){
         this.currentDiscount=newCurrentDiscount;
     }
